requirements.txt
Jinja2==3.1.2
dnspython==2.2.1
requests==2.32.5
mysql-connector-python==8.0.33